Backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-dotenv
pydantic[email]